requirements.txt
langchain 
wikipedia 
langchain-community 
langchain-openai
langchain-anthropic
langchain-huggingface
python-dotenv
pydantic
ddgs
pyowm
faiss-cpu
sentence-transformers